% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{captcha_classify}
\alias{captcha_classify}
\title{Classify CAPTCHAS}
\usage{
captcha_classify(files, folder, remove = TRUE)
}
\arguments{
\item{files}{character vector containing file paths}

\item{folder}{folder to save new files.}

\item{remove}{if true, remove original file. Otherwise, copy it.}
}
\description{
Classify CAPTCHAS
}
