% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{captcha_download}
\alias{captcha_download}
\title{Download n captcha images from TRT3's website at once}
\usage{
captcha_download(n, repeat_captcha_n_times = 1, dir, sleep = 1)
}
\arguments{
\item{n}{number of captchas to download}

\item{repeat_captcha_n_times}{number of times that one sequence of letters should be repeated (varying just the noise draws between them)}

\item{dir}{character string naming a directory path for writing the image}

\item{sleep}{time in seconds for the system wait until next try if the attempt to download fails}

\item{repeat_last_captcha}{boolean indicating if the letters of the last captcha download should persist, changing only the noise draws.}
}
\value{
list of responses
}
\description{
Download n captcha images from TRT3's website at once
}
