% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{captcha_download_one}
\alias{captcha_download_one}
\title{Download one captcha image from TRT3's website}
\usage{
captcha_download_one(dir, sleep = 1, repeat_last_captcha = FALSE)
}
\arguments{
\item{dir}{character string naming a directory path for writing the image}

\item{sleep}{time in seconds for the system wait until next try if the attempt to download fails}

\item{repeat_last_captcha}{boolean indicating if the letters of the last captcha download should persist, changing only the noise draws.}
}
\value{
response
}
\description{
Download just one captcha image from TRT3's website
}
