% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraculo.R
\name{oraculo}
\alias{oraculo}
\title{Tells whether the model broke the captcha}
\usage{
oraculo(model = NULL)
}
\arguments{
\item{model}{model to predict captcha.}
}
\value{
logical
}
\description{
Returns TRUE if the model managed to break the captcha and FALSE otherwise.
  If model is NULL, we ask the user to insert the captcha.
}
