% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oraculo.R
\name{oraculo}
\alias{oraculo}
\title{Tells whether the model passed CAPTCHA}
\usage{
oraculo(model = NULL)
}
\arguments{
\item{model}{model to predict CAPTCHA.}
}
\value{
logical
}
\description{
Returns TRUE if the model bypass the CAPTCHA and FALSE otherwise.
  If model is NULL, we ask the user to type the text manually.
}
